Expedientes V5

+++ models/EntidadSerie.php
@@ -34,14 +34,17 @@ class EntidadSerie extends Model
         ];
     }
 
-    protected function afterDelete()
+    protected function afterUpdate()
     {
-        $PermisoSerie = PermisoSerie::findAllByAttributes(['fk_entidad_serie' => $this->getPK()]);
-        if ($PermisoSerie) {
-            foreach ($PermisoSerie as $instance) {
-                $instance->delete();
+        if (!$this->estado) {
+            $PermisoSerie = PermisoSerie::findAllByAttributes(['fk_entidad_serie' => $this->getPK()]);
+            if ($PermisoSerie) {
+                foreach ($PermisoSerie as $instance) {
+                    $instance->delete();
+                }
             }
         }
+
         return true;
     }
 
@@ -67,30 +70,45 @@ class EntidadSerie extends Model
                             $codPadreExp = $consPadre[0]['idexpediente'];
                         }
                     }
-
-                    $Expediente = new Expediente();
-                    $attributes = [
-                        'nombre' => $Serie->nombre,
-                        'fondo' => $Serie->nombre,
-                        'descripcion' => $Serie->nombre,
-                        'codigo' => $Serie->codigo,
-                        'codigo_numero' => $Serie->codigo,
-                        'cod_padre' => $codPadreExp,
-                        'fk_idcaja' => 0,
-                        'propietario' => 0,
-                        'fk_serie' => $this->fk_serie,
-                        'fk_dependencia' => $this->fk_dependencia,
-                        'cod_arbol' => 0,
-                        'agrupador' => 2,
-                        'fk_entidad_serie' => $this->identidad_serie,
-                        'nucleo' => 1
-                    ];
-                    $Expediente->SetAttributes($attributes);
-                    $infoExpediente = $Expediente->CreateExpediente();
-                    if ($infoExpediente['exito']) {
-                        $response['data']['idexpediente'] = $Expediente->getPK();
+                    $existExp = busca_filtro_tabla("idexpediente,estado", "expediente", "fk_dependencia={$this->fk_dependencia} and fk_serie={$this->fk_serie} and nucleo=1 and estado=1 and agrupador=2", "", $conn);
+                    if (!$existExp['numcampos']) {
+                        $Expediente = new Expediente();
+                        $attributes = [
+                            'nombre' => $Serie->nombre,
+                            'fondo' => $Serie->nombre,
+                            'descripcion' => $Serie->nombre,
+                            'codigo' => $Serie->codigo,
+                            'codigo_numero' => $Serie->codigo,
+                            'cod_padre' => $codPadreExp,
+                            'fk_idcaja' => 0,
+                            'propietario' => 0,
+                            'fk_serie' => $this->fk_serie,
+                            'fk_dependencia' => $this->fk_dependencia,
+                            'cod_arbol' => 0,
+                            'agrupador' => 2,
+                            'fk_entidad_serie' => $this->identidad_serie,
+                            'nucleo' => 1
+                        ];
+                        $Expediente->SetAttributes($attributes);
+                        $infoExpediente = $Expediente->CreateExpediente();
+                        if ($infoExpediente['exito']) {
+                            $response['data']['idexpediente'] = $Expediente->getPK();
+                            $response['exito'] = 1;
+                        }
+                    } else {
+                        if (!$existExp[0]['estado']) {
+                            $Expediente = new Expediente($existExp[0]['idexpediente']);
+                            $attributes = [
+                                'estado' => 0,
+                                'fecha_eliminacion' => date('Y-m-d H:i:s')
+                            ];
+                            $Expediente->SetAttributes($attributes);
+                            $Expediente->update();
+                        }
+                        $response['data']['idexpediente'] = $existExp[0]['idexpediente'];
                         $response['exito'] = 1;
                     }
+
                 } else {
                     $response['message'] = 'NO se encontro la serie a vincular';
                 }
@@ -111,25 +129,11 @@ class EntidadSerie extends Model
             'exito' => 0,
             'message' => '',
         ];
-        $expeVinc = $this->getExpedienteFk();
-        if ($expeVinc) {
-            $attributes = [
-                'estado' => 0,
-                'fecha_eliminacion' => date('Y-m-d H:i:s')
-            ];
-            foreach ($expeVinc as $instance) {
-                $instance->SetAttributes($attributes);
-                $instance->update();
-            }
-            $response['exito'] = 1;
-        } else {
-            var_dump($this->delete());
-            die("--a--");
-            if ($this->delete()) {
+        $this->estado = 0;
+        $this->fecha_eliminacion = date('Y-m-d H:i:s');
+        $this->update();
 
-            }
-            $response['exito'] = 1;
-        }
+        $response['exito'] = 1;
         return $response;
     }
 
@@ -143,84 +147,51 @@ class EntidadSerie extends Model
         $Dependencia = $this->getDependenciaFk();
         if ($Dependencia) {
             $Dependencia = $Dependencia[0];
-            $Serie = $this->getSerieFk();
-            if ($Serie) {
-                $Serie = $Serie[0];
-                if ($Serie->tipo != 3) {
-                    $hijosSerie = busca_filtro_tabla("idserie", "serie", "cod_padre=" . $this->fk_serie, "", $conn);
-                    if ($hijosSerie['numcampos']) {
-                        for ($i = 0; $i < $hijosSerie['numcampos']; $i++) {
-                            $fksDependencias = busca_filtro_tabla("distinct fk_dependencia", "entidad_serie", "fk_serie={$hijosSerie[$i]['idserie']} and estado=1", "", $conn);
-                            if ($fksDependencias['numcampos']) {
-                                for ($j = 0; $j < $fksDependencias['numcampos']; $j++) {
-                                    $existEntSer = busca_filtro_tabla("identidad_serie", "entidad_serie", "fk_serie={$this->fk_serie} and fk_dependencia={$fksDependencias[$j]['fk_dependencia']} and estado=1", "", $conn);
-                                    if (!$existEntSer['numcampos']) {
-                                        $attributes = [
-                                            'fk_serie' => $this->fk_serie,
-                                            'fk_dependencia' => $fksDependencias[$j]['fk_dependencia'],
-                                            'estado' => 1,
-                                            'fecha_creacion' => date('Y-m-d H:i:s')
-                                        ];
-                                        $newEntSerie = new self();
-                                        $newEntSerie->SetAttributes($attributes);
-                                        $ok = $newEntSerie->CreateEntidadSerie();
-                                        if (!$ok['exito']) {
-                                            file_put_contents('logEntidadSerie.txt', print_r($ok, true));
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
 
-                $idsDep = explode('.', $Dependencia->codigo_arbol);
-                $idsExp = [];
-                foreach ($idsDep as $key => $idDependencia) {
-                    $existDep = busca_filtro_tabla("idexpediente", "expediente", "fk_dependencia={$idDependencia} and nucleo=1 and estado=1", "", $conn);
-                    if ($existDep['numcampos']) {
-                        $idsExp[$key] = $existDep[0]['idexpediente'];
-                    } else {
-                        $DependenciaData = new Dependencia($idDependencia);
-                        if ($key == 0) {
-                            $codPadreExp = $key;
-                        } else {
-                            $codPadreExp = $idsExp[$key - 1];
-                        }
-                        $Expediente = new Expediente();
-                        $attributes = [
-                            'nombre' => $DependenciaData->nombre,
-                            'fondo' => $DependenciaData->nombre,
-                            'descripcion' => $DependenciaData->nombre,
-                            'codigo' => $DependenciaData->codigo,
-                            'codigo_numero' => $DependenciaData->codigo,
-                            'cod_padre' => $codPadreExp,
-                            'fk_idcaja' => 0,
-                            'propietario' => 0,
-                            'fk_serie' => 0,
-                            'fk_dependencia' => $idDependencia,
-                            'cod_arbol' => 0,
-                            'agrupador' => 1,
-                            'fk_entidad_serie' => 0,
-                            'nucleo' => 1
-                        ];
-                        $Expediente->SetAttributes($attributes);
-                        $infoExpediente = $Expediente->CreateExpediente();
-                        if ($infoExpediente['exito']) {
-                            $idsExp[$key] = $Expediente->getPK();
-                        }
-                    }
-                }
-                if (count($idsDep) == count($idsExp)) {
-                    $response['exito'] = 1;
-                    $response['data']['idexpediente'] = $idsExp;
+            $idsDep = explode('.', $Dependencia->codigo_arbol);
+            $idsExp = [];
+            foreach ($idsDep as $key => $idDependencia) {
+                //TODO: El estado de expediente solo debe cambiarse cuando se elimine la serie que hace relacion
+                $existDep = busca_filtro_tabla("idexpediente", "expediente", "fk_dependencia={$idDependencia} and nucleo=1 and estado=1 and agrupador=1", "", $conn);
+                if ($existDep['numcampos']) {
+                    $idsExp[$key] = $existDep[0]['idexpediente'];
                 } else {
-                    $response['message'] = 'Error, No se crearon todos los expedientes';
+                    $DependenciaData = new Dependencia($idDependencia);
+                    if ($key == 0) {
+                        $codPadreExp = $key;
+                    } else {
+                        $codPadreExp = $idsExp[$key - 1];
+                    }
+                    $Expediente = new Expediente();
+                    $attributes = [
+                        'nombre' => $DependenciaData->nombre,
+                        'fondo' => $DependenciaData->nombre,
+                        'descripcion' => $DependenciaData->nombre,
+                        'codigo' => $DependenciaData->codigo,
+                        'codigo_numero' => $DependenciaData->codigo,
+                        'cod_padre' => $codPadreExp,
+                        'fk_idcaja' => 0,
+                        'propietario' => 0,
+                        'fk_serie' => 0,
+                        'fk_dependencia' => $idDependencia,
+                        'cod_arbol' => 0,
+                        'agrupador' => 1,
+                        'fk_entidad_serie' => 0,
+                        'nucleo' => 1
+                    ];
+                    $Expediente->SetAttributes($attributes);
+                    $infoExpediente = $Expediente->CreateExpediente();
+                    if ($infoExpediente['exito']) {
+                        $idsExp[$key] = $Expediente->getPK();
+                    }
                 }
-            } else {
-                $response['message'] = 'Error, la serie NO existe! id:' . $this->fk_serie;
             }
-
+            if (count($idsDep) == count($idsExp)) {
+                $response['exito'] = 1;
+                $response['data']['idexpediente'] = $idsExp;
+            } else {
+                $response['message'] = 'Error, No se crearon todos los expedientes';
+            }
         } else {
             $response['message'] = 'Error, la dependencia NO existe! id:' . $this->fk_dependencia;
         }
